feat: Add confirm modal before deleting project
